requirements.txt
openai==0.27.8
langchain==0.0.195
pypdf
pinecone-io
pinecone-client
pinecone-client[grpc]
openai
unstructured
pdf2image
dnspython
loguru
numpy
python-dateutil
pyyaml
requests
tqdm
typing_extensions
urllib3
tiktoken
pdfminer.six
